image cache use new cache code

+++ main/java/model/cache/ImageCacheManager.java
@@ -8,36 +8,31 @@ import java.util.Map;
  * Created by xlo on 15-9-16.
  * it's the image cache manager
  */
-public class ImageCacheManager extends CacheManager {
+public class ImageCacheManager {
     protected Map<String, File> fileMap;
 
-    private static ImageCacheManager imageCacheManager;
+    private static WaitDeleteCacheManager waitDeleteCacheManager;
 
     private ImageCacheManager() {
         this.fileMap = new HashMap<>();
     }
 
-    public static ImageCacheManager getImageCacheManager() {
-        if (imageCacheManager == null) {
+    public static WaitDeleteCacheManager getImageCacheManager() {
+        if (waitDeleteCacheManager == null) {
             synchronized (ImageCacheManager.class) {
-                if (imageCacheManager == null) {
-                    imageCacheManager = new ImageCacheManager();
+                if (waitDeleteCacheManager == null) {
+                    waitDeleteCacheManager = new WaitDeleteCacheManager(5000, (object) -> {
+                        File file = new File(object.toString());
+                        while (file.exists() && !file.delete()) {
+                            try {
+                                Thread.sleep(500);
+                            } catch (InterruptedException ignored) {
+                            }
+                        }
+                    });
                 }
             }
         }
-        return imageCacheManager;
-    }
-
-    @Override
-    protected CacheObject newCacheObject(Object object) {
-        return new WaitDeleteCacheObject(object, () -> {
-            File file = new File(object.toString());
-            while (file.exists() && !file.delete()) {
-                try {
-                    Thread.sleep(500);
-                } catch (InterruptedException ignored) {
-                }
-            }
-        });
+        return waitDeleteCacheManager;
     }
 }
